requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
blis==0.4.1
boto3==1.14.63
botocore==1.17.63
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
cycler==0.10.0
cymem==2.0.3
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.2
gensim==3.8.3
idna==2.10
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.2
olefile==0.46
pandas==1.1.2
pdfminer.six==20181108
Pillow==7.2.0
plac==1.1.3
plotly==4.10.0
preshed==3.0.2
pycryptodome==3.9.8
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
python-pptx==0.6.18
pytz==2020.1
regex==2020.7.14
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.12.0
smart-open==2.2.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
SpeechRecognition==3.8.1
srsly==1.0.2
texthero==1.0.9
textract==1.6.3
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.49.0
tzlocal==1.5.1
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
wordcloud==1.8.0
xlrd==1.2.0
XlsxWriter==1.3.6
https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-2.3.1/es_core_news_lg-2.3.1.tar.gz#egg=en_core_web_sm
neomodel~=4.0.1